[background-task] Address review: fix spacing, rename DevClient, drop await
- Fix spacing convention in tsconfig.json and expo-module.config.json
- Remove @expo/devtools peerDep (available via expo)
- Rename build-cli -> build:cli script
- Rename BackgroundTaskDevClient -> BackgroundTaskDevToolsClient
- Remove await on maybeInitDevToolsAsync (non-blocking)
- Remove cli/package.json, rename index.ts -> index.mts
- Update entryPoint to cli/build/index.mjs
