+++ packages/expo-background-task/cli/tsconfig.json
@@ -4,6 +4,6 @@
     "outDir": "build",
     "rootDir": "src"
   },
-  "include": [ "./src" ],
-  "exclude": [ "**/__mocks__/*", "**/__tests__/*" ]
+  "include": ["./src"],
+  "exclude": ["**/__mocks__/*", "**/__tests__/*"]
 }
